Refactor: move IconButton.tsx to /IconsButton
